cant delete product

+++ src/main/java/com/fox/factory/service/CommentService.java
@@ -1,19 +1,27 @@
 package com.fox.factory.service;
 
+import com.fox.factory.entities.Comment;
 import com.fox.factory.entities.dto.CommentCreateDto;
 import com.fox.factory.entities.dto.CommentsDto;
 import com.fox.factory.entities.dto.SubCommentDto;
 import com.fox.factory.entities.dto.UserInCommentDto;
 import com.fox.factory.repositories.CommentRepository;
+import com.fox.factory.security.JwtAuthenticationFilter;
+import com.fox.factory.security.Role;
 import com.fox.factory.service.mappers.CommentMapper;
 import com.fox.factory.service.mappers.UserMapper;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Objects;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
@@ -23,6 +31,7 @@ public class CommentService {
     private final CommentRepository repository;
     private final CommentMapper mapper;
     private final UserMapper userMapper;
+    private final JwtAuthenticationFilter jwtAuthenticationFilter;
     @Transactional
     public CommentsDto create(CommentCreateDto a){
         return mapper.toDto(repository.save(mapper.toEntityCreate(a)));
@@ -55,9 +64,11 @@ public class CommentService {
 
     }
     @Transactional
-    public CommentsDto addSubComment(Long id, SubCommentDto sub){
+    public CommentsDto addSubComment(HttpServletRequest req,  Long id, SubCommentDto sub){
+        var user = jwtAuthenticationFilter.get_user_from_req(req);
         var commen = mapper.fromSubCommentDto(sub);
         commen.setPublishDate(LocalDate.now());
+        commen.setAuthor(user);
         final var a = repository.save(commen);
         return repository.findById(id)
                 .map(comment -> comment.addToSubComments(a))
@@ -75,7 +86,19 @@ public class CommentService {
                 .orElse(null);
     }
     @Transactional
-    public void delete(Long id){
+    public void delete(HttpServletRequest req,  Long id){
+        var commnent = repository.findById(id);
+        if (commnent.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No comment");
+        }
+        var c = commnent.get();
+
+        var user = jwtAuthenticationFilter.get_user_from_req(req);
+
+        if (!Objects.equals(c.getAuthor().getId(), user.getId()) && user.getRole() == Role.USER){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         repository.deleteById(id);
     }
 
@@ -84,5 +107,30 @@ public class CommentService {
     @Transactional
     public CommentsDto findById(Long id) {
         return mapper.toDto(repository.findById(id).orElse(null));
+    }
+
+    public void delete_sub(HttpServletRequest req, Long id, Long subId) {
+        var commnent = repository.findById(id);
+        var sub_comment = repository.findById(subId);
+
+        if (commnent.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No comment");
+        }
+        var c = commnent.get();
+
+        if (sub_comment.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No comment");
+        }
+        var cc = sub_comment.get();
+
+        var user = jwtAuthenticationFilter.get_user_from_req(req);
+
+        if (!Objects.equals(cc.getAuthor().getId(), user.getId()) && user.getRole() == Role.USER){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
+        c.setSubComments(c.getSubComments().stream().filter(comment -> !Objects.equals(comment.getId(), cc.getId())).collect(Collectors.toSet()));
+        repository.save(c);
+        repository.delete(cc);
     }
 }
